requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-environ==0.12.0
psycopg2-binary==2.9.10
